+++ app/controllers/foreman_azure_rm/concerns/compute_resources_controller_extensions.rb
@@ -10,9 +10,15 @@ module ForemanAzureRM
         render :available_resource_groups, :layout => 'api/v2/layouts/index_layout'
       end
 
+      def available_sizes
+        compute_resource = ComputeResource.find_by_id(params[:id])
+        @available_sizes = compute_resource.vm_sizes(params[:region_id])
+        render :available_sizes, :layout => 'api/v2/layouts/index_layout'
+      end
+
       def action_permission
         case params[:action]
-          when 'available_resource_groups'
+          when 'available_resource_groups', 'available_sizes'
             :view
           else
             super