+++ app/controllers/foreman_azure_rm/concerns/hosts_controller_extensions.rb
@@ -5,7 +5,7 @@ module ForemanAzureRM
       def sizes
         if (azure_rm_resource = Image.unscoped.find_by_uuid(params[:image_id])).present?
           resource = azure_rm_resource.compute_resource
-          render :json => resource.vm_sizes(params[:location_string])
+          render :json => resource.vm_sizes(params[:location_string]).map { |size| size.name }
         else
           no_sizes = _('The location you selected has no sizes associated with it')
           render :json => "[\"#{no_sizes}\"]"
@@ -16,7 +16,7 @@ module ForemanAzureRM
         azure_rm_image = Image.unscoped.find_by_uuid(params[:image_id])
         if azure_rm_image.present?
           azure_rm_resource = azure_rm_image.compute_resource
-          subnets = azure_rm_resource.subnets(params[:location])
+          subnets           = azure_rm_resource.subnets(params[:location])
           if subnets.present?
             render :json => azure_rm_resource.subnets(params[:location])
           else