+++ config/routes.rb
@@ -9,6 +9,7 @@ Rails.application.routes.draw do
     scope "(:apiv)", :module => :v2, :defaults => { :apiv => 'v2' }, :apiv => /v1|v2/, :constraints => ApiConstraints.new(:version => 2, :default => true) do
       resources :compute_resources, :except => [:new, :edit] do
         get :available_resource_groups, :on => :member
+        get '(:region_id)/available_sizes', :to => 'compute_resources#available_sizes', :on => :member
       end
     end
   end